+++ awx/main/tasks/receptor.py
@@ -411,9 +411,9 @@ class AWXReceptorJob:
             streamer='process',
             quiet=True,
             _input=resultfile,
-            event_handler=self.task.event_handler,
-            finished_callback=self.task.finished_callback,
-            status_handler=self.task.status_handler,
+            event_handler=self.task.runner_callback.event_handler,
+            finished_callback=self.task.runner_callback.finished_callback,
+            status_handler=self.task.runner_callback.status_handler,
             **self.runner_params,
         )
 
@@ -458,7 +458,7 @@ class AWXReceptorJob:
             if processor_future.done():
                 return processor_future.result()
 
-            if self.task.cancel_callback():
+            if self.task.runner_callback.cancel_callback():
                 result = namedtuple('result', ['status', 'rc'])
                 return result('canceled', 1)
